refactor(project1): add test name

+++ Project1/test/test.sh
@@ -1,8 +1,12 @@
 #!/bin/sh
 progTmp="$(mktemp -d)"
 sysTmp="$(mktemp -d)"
-testName="countryCity"
+testName="$1"
 cd "$(dirname "$0")/" || exit 1
+if [ ! -f "$testName.sh" ]; then
+    echo "Test suite not found!"
+    exit 1
+fi
 sh -c 'cd .. && make'
 cp -a "${testName}.sh" "${testName}"/* "${progTmp}/"
 cp -a "${testName}.sh" "${testName}"/* "${sysTmp}/"
